#117 New flat icon (based on old icon) (#118)
* #117 New flat icon (based on old icon)

* Added logo-small.svg

Sorry for don't add it in previous commit, but for any reason my GH Desktop Client ignored it.

* Update logo-small.svg

* Update frontend.less 

Added class navbar-header > a > img. Now the <img> size need be declarated in CSS, not in the file, because now is used dynamic SVG file instead of a static PNG image.

* Added icon.ico

This will be used when compile PO into an app to any platform.